Updating package structure
